+++ src/backends/s3.rs
@@ -144,7 +144,7 @@ pub struct UpdateBuilder {
     region: Option<String>,
     bin_name: Option<String>,
     bin_install_path: Option<PathBuf>,
-    bin_path_in_archive: Option<PathBuf>,
+    bin_path_in_archive: Option<String>,
     show_download_progress: bool,
     show_output: bool,
     no_confirm: bool,
@@ -153,6 +153,8 @@ pub struct UpdateBuilder {
     progress_template: String,
     progress_chars: String,
     auth_token: Option<String>,
+    #[cfg(feature = "signatures")]
+    verifying_keys: Vec<[u8; zipsign_api::PUBLIC_KEY_LENGTH]>,
 }
 
 impl Default for UpdateBuilder {
@@ -174,6 +176,8 @@ impl Default for UpdateBuilder {
             progress_template: DEFAULT_PROGRESS_TEMPLATE.to_string(),
             progress_chars: DEFAULT_PROGRESS_CHARS.to_string(),
             auth_token: None,
+            #[cfg(feature = "signatures")]
+            verifying_keys: vec![],
         }
     }
 }
@@ -237,10 +241,10 @@ impl UpdateBuilder {
     /// Set the exe's name. Also sets `bin_path_in_archive` if it hasn't already been set.
     pub fn bin_name(&mut self, name: &str) -> &mut Self {
         let raw_bin_name = format!("{}{}", name.trim_end_matches(EXE_SUFFIX), EXE_SUFFIX);
-        self.bin_name = Some(raw_bin_name.clone());
         if self.bin_path_in_archive.is_none() {
-            self.bin_path_in_archive = Some(PathBuf::from(raw_bin_name));
+            self.bin_path_in_archive = Some(raw_bin_name.to_owned());
         }
+        self.bin_name = Some(raw_bin_name);
         self
     }
 
@@ -258,13 +262,20 @@ impl UpdateBuilder {
     /// the path to the binary (from the root of the tarball) is not equal to just
     /// the `bin_name`.
     ///
+    /// This also supports variable paths:
+    /// - `{{ bin }}` is replaced with the value of `bin_name`
+    /// - `{{ target }}` is replaced with the value of `target`
+    /// - `{{ version }}` is replaced with the value of `target_version` if set,
+    /// otherwise the value of the latest available release version is used.
+    ///
     /// # Example
     ///
-    /// For a tarball `myapp.tar.gz` with the contents:
+    /// For a `myapp` binary with `windows` target and latest release version `1.2.3`,
+    /// the tarball `myapp.tar.gz` has the contents:
     ///
     /// ```shell
     /// myapp.tar/
-    ///  |------- bin/
+    ///  |------- windows-1.2.3-bin/
     ///  |         |--- myapp  # <-- executable
     /// ```
     ///
@@ -274,13 +285,13 @@ impl UpdateBuilder {
     /// # use self_update::backends::s3::Update;
     /// # fn run() -> Result<(), Box<::std::error::Error>> {
     /// Update::configure()
-    ///     .bin_path_in_archive("bin/myapp")
+    ///     .bin_path_in_archive("{{ target }}-{{ version }}-bin/{{ bin }}")
     /// #   .build()?;
     /// # Ok(())
     /// # }
     /// ```
     pub fn bin_path_in_archive(&mut self, bin_path: &str) -> &mut Self {
-        self.bin_path_in_archive = Some(PathBuf::from(bin_path));
+        self.bin_path_in_archive = Some(bin_path.to_owned());
         self
     }
 
@@ -318,6 +329,19 @@ impl UpdateBuilder {
         self
     }
 
+    /// Specify a slice of ed25519ph verifying keys to validate a download's authenticy
+    ///
+    /// If the feature is activated AND at least one key was provided, a download is verifying.
+    /// At least one key has to match.
+    #[cfg(feature = "signatures")]
+    pub fn verifying_keys(
+        &mut self,
+        keys: impl Into<Vec<[u8; zipsign_api::PUBLIC_KEY_LENGTH]>>,
+    ) -> &mut Self {
+        self.verifying_keys = keys.into();
+        self
+    }
+
     /// Confirm config and create a ready-to-use `Update`
     ///
     /// * Errors:
@@ -349,8 +373,8 @@ impl UpdateBuilder {
                 bail!(Error::Config, "`bin_name` required")
             },
             bin_install_path,
-            bin_path_in_archive: if let Some(ref path) = self.bin_path_in_archive {
-                path.to_owned()
+            bin_path_in_archive: if let Some(ref bin_path) = self.bin_path_in_archive {
+                bin_path.to_owned()
             } else {
                 bail!(Error::Config, "`bin_path_in_archive` required")
             },
@@ -366,6 +390,8 @@ impl UpdateBuilder {
             show_output: self.show_output,
             no_confirm: self.no_confirm,
             auth_token: self.auth_token.clone(),
+            #[cfg(feature = "signatures")]
+            verifying_keys: self.verifying_keys.clone(),
         }))
     }
 }
@@ -382,13 +408,15 @@ pub struct Update {
     target_version: Option<String>,
     bin_name: String,
     bin_install_path: PathBuf,
-    bin_path_in_archive: PathBuf,
+    bin_path_in_archive: String,
     show_download_progress: bool,
     show_output: bool,
     no_confirm: bool,
     progress_template: String,
     progress_chars: String,
     auth_token: Option<String>,
+    #[cfg(feature = "signatures")]
+    verifying_keys: Vec<[u8; zipsign_api::PUBLIC_KEY_LENGTH]>,
 }
 
 impl Update {
@@ -466,7 +494,7 @@ impl ReleaseUpdate for Update {
         self.bin_install_path.clone()
     }
 
-    fn bin_path_in_archive(&self) -> PathBuf {
+    fn bin_path_in_archive(&self) -> String {
         self.bin_path_in_archive.clone()
     }
 
@@ -492,6 +520,11 @@ impl ReleaseUpdate for Update {
 
     fn auth_token(&self) -> Option<String> {
         self.auth_token.clone()
+    }
+
+    #[cfg(feature = "signatures")]
+    fn verifying_keys(&self) -> &[[u8; zipsign_api::PUBLIC_KEY_LENGTH]] {
+        &self.verifying_keys
     }
 }