+++ src/update.rs
@@ -1,10 +1,11 @@
+use regex::Regex;
 use reqwest::{self, header};
+use std::borrow::Cow;
+use std::env::consts::{ARCH, OS};
 use std::fs;
-#[cfg(not(windows))]
-use std::os::unix::fs::PermissionsExt;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
-use crate::{confirm, errors::*, version, Download, Extract, Move, Status};
+use crate::{confirm, errors::*, version, Download, Extract, Status};
 
 /// Release asset information
 #[derive(Clone, Debug, Default)]
@@ -64,11 +65,12 @@ impl Release {
             .iter()
             .find(|asset| {
                 asset.name.contains(target)
-                    && if let Some(i) = identifier {
-                        asset.name.contains(i)
-                    } else {
-                        true
-                    }
+                    || (asset.name.contains(OS) && asset.name.contains(ARCH))
+                        && if let Some(i) = identifier {
+                            asset.name.contains(i)
+                        } else {
+                            true
+                        }
             })
             .cloned()
     }
@@ -103,7 +105,7 @@ pub trait ReleaseUpdate {
     fn bin_install_path(&self) -> PathBuf;
 
     /// Path of the binary to be extracted from release package
-    fn bin_path_in_archive(&self) -> PathBuf;
+    fn bin_path_in_archive(&self) -> String;
 
     /// Flag indicating if progress information shall be output when downloading a release
     fn show_download_progress(&self) -> bool;
@@ -122,6 +124,12 @@ pub trait ReleaseUpdate {
 
     /// Authorisation token for communicating with backend
     fn auth_token(&self) -> Option<String>;
+
+    /// ed25519ph verifying keys to validate a download's authenticy
+    #[cfg(feature = "signatures")]
+    fn verifying_keys(&self) -> &[[u8; zipsign_api::PUBLIC_KEY_LENGTH]] {
+        &[]
+    }
 
     /// Construct a header with an authorisation entry if an auth token is provided
     fn api_headers(&self, auth_token: &Option<String>) -> Result<header::HeaderMap> {
@@ -151,25 +159,6 @@ pub trait ReleaseUpdate {
     fn update_extended(&self) -> Result<UpdateStatus> {
         let bin_install_path = self.bin_install_path();
         let bin_name = self.bin_name();
-
-        let tmp_dir_parent = bin_install_path
-            .parent()
-            .map(PathBuf::from)
-            .ok_or_else(|| Error::Update("Failed to determine parent dir".into()))?;
-        let tmp_backup_dir_prefix = format!("__{}_backup", bin_name);
-        let tmp_backup_filename = tmp_backup_dir_prefix.clone();
-
-        if cfg!(windows) {
-            // Windows executables can not be removed while they are running, which prevents clean up
-            // of the temporary directory by the `tempfile` crate after we move the running executable
-            // into it during an update. We clean up any previously created temporary directories here.
-            // Ignore errors during cleanup since this is not critical for completing the update.
-            let _ = cleanup_backup_temp_directories(
-                &tmp_dir_parent,
-                &tmp_backup_dir_prefix,
-                &tmp_backup_filename,
-            );
-        }
 
         let current_version = self.current_version();
         let target = self.target();
@@ -235,10 +224,7 @@ pub trait ReleaseUpdate {
             confirm("Do you want to continue? [Y/n] ")?;
         }
 
-        let tmp_archive_dir_prefix = format!("{}_download", bin_name);
-        let tmp_archive_dir = tempfile::Builder::new()
-            .prefix(&tmp_archive_dir_prefix)
-            .tempdir_in(&tmp_dir_parent)?;
+        let tmp_archive_dir = tempfile::TempDir::new()?;
         let tmp_archive_path = tmp_archive_dir.path().join(&target_asset.name);
         let mut tmp_archive = fs::File::create(&tmp_archive_path)?;
 
@@ -254,33 +240,36 @@ pub trait ReleaseUpdate {
 
         download.download_to(&mut tmp_archive)?;
 
-        print_flush(show_output, "Extracting archive... ")?;
-        let bin_path_in_archive = self.bin_path_in_archive();
-        Extract::from_source(&tmp_archive_path)
-            .extract_file(tmp_archive_dir.path(), &bin_path_in_archive)?;
-        let new_exe = tmp_archive_dir.path().join(&bin_path_in_archive);
+        #[cfg(feature = "signatures")]
+        verify_signature(&tmp_archive_path, self.verifying_keys())?;
 
-        // Make executable
-        #[cfg(not(windows))]
-        {
-            let mut permissions = fs::metadata(&new_exe)?.permissions();
-            permissions.set_mode(0o755);
-            fs::set_permissions(&new_exe, permissions)?;
+        print_flush(show_output, "Extracting archive... ")?;
+
+        let bin_path_str = Cow::Owned(self.bin_path_in_archive());
+
+        /// Substitute the `var` variable in a string with the given `val` value.
+        ///
+        /// Variable format: `{{ var }}`
+        fn substitute<'a: 'b, 'b>(str: &'a str, var: &str, val: &str) -> Cow<'b, str> {
+            let format = format!(r"\{{\{{[[:space:]]*{}[[:space:]]*\}}\}}", var);
+            Regex::new(&format).unwrap().replace_all(str, val)
         }
+
+        let bin_path_str = substitute(&bin_path_str, "version", &release.version);
+        let bin_path_str = substitute(&bin_path_str, "target", &target);
+        let bin_path_str = substitute(&bin_path_str, "bin", &bin_name);
+        let bin_path_str = bin_path_str.as_ref();
+
+        Extract::from_source(&tmp_archive_path)
+            .extract_file(tmp_archive_dir.path(), bin_path_str)?;
+        let new_exe = tmp_archive_dir.path().join(bin_path_str);
 
         println(show_output, "Done");
 
         print_flush(show_output, "Replacing binary file... ")?;
-
-        let tmp_backup_dir = tempfile::Builder::new()
-            .prefix(&tmp_backup_dir_prefix)
-            .tempdir_in(&tmp_dir_parent)?;
-        let tmp_file_path = tmp_backup_dir.path().join(&tmp_backup_filename);
-
-        Move::from_source(&new_exe)
-            .replace_using_temp(&tmp_file_path)
-            .to_dest(&bin_install_path)?;
+        self_replace::self_replace(new_exe)?;
         println(show_output, "Done");
+
         Ok(UpdateStatus::Updated(release))
     }
 }
@@ -300,35 +289,47 @@ fn println(show_output: bool, msg: &str) {
     }
 }
 
-fn cleanup_backup_temp_directories<P: AsRef<Path>>(
-    tmp_dir_parent: P,
-    tmp_dir_prefix: &str,
-    expected_tmp_filename: &str,
-) -> Result<()> {
-    for entry in fs::read_dir(tmp_dir_parent)? {
-        let entry = entry?;
-        let tmp_dir_name = if let Ok(tmp_dir_name) = entry.file_name().into_string() {
-            tmp_dir_name
-        } else {
-            continue;
-        };
+#[cfg(feature = "signatures")]
+fn verify_signature(
+    archive_path: &std::path::Path,
+    keys: &[[u8; zipsign_api::PUBLIC_KEY_LENGTH]],
+) -> crate::Result<()> {
+    if keys.is_empty() {
+        return Ok(());
+    }
 
-        // For safety, check that the temporary directory contains only the expected backup
-        // binary file before removing. If subdirectories or other files exist then the user
-        // is using the temp directory for something else. This is unlikely, but we should
-        // be careful with `fs::remove_dir_all`.
-        let is_expected_tmp_file = |tmp_file_entry: std::io::Result<fs::DirEntry>| {
-            tmp_file_entry
-                .ok()
-                .filter(|e| e.file_name() == expected_tmp_filename)
-                .is_some()
-        };
+    println!("Verifying downloaded file...");
 
-        if tmp_dir_name.starts_with(tmp_dir_prefix)
-            && fs::read_dir(entry.path())?.all(is_expected_tmp_file)
-        {
-            fs::remove_dir_all(entry.path())?;
+    let archive_kind = crate::detect_archive(archive_path)?;
+    #[cfg(any(feature = "archive-tar", feature = "archive-zip"))]
+    {
+        let context = archive_path
+            .file_name()
+            .and_then(|s| s.to_str())
+            .map(|s| s.as_bytes())
+            .ok_or(Error::NonUTF8)?;
+
+        let keys = keys.iter().copied().map(Ok);
+        let keys =
+            zipsign_api::verify::collect_keys(keys).map_err(zipsign_api::ZipsignError::from)?;
+
+        let mut exe = std::fs::File::open(archive_path)?;
+
+        match archive_kind {
+            #[cfg(feature = "archive-tar")]
+            crate::ArchiveKind::Tar(Some(crate::Compression::Gz)) => {
+                zipsign_api::verify::verify_tar(&mut exe, &keys, Some(context))
+                    .map_err(zipsign_api::ZipsignError::from)?;
+                return Ok(());
+            }
+            #[cfg(feature = "archive-zip")]
+            crate::ArchiveKind::Zip => {
+                zipsign_api::verify::verify_zip(&mut exe, &keys, Some(context))
+                    .map_err(zipsign_api::ZipsignError::from)?;
+                return Ok(());
+            }
+            _ => {}
         }
     }
-    Ok(())
+    Err(Error::NoSignatures(archive_kind))
 }